+++ tilings/strategies/factor.py
@@ -42,7 +42,10 @@ class FactorStrategy(CartesianProductStrategy[Tiling, GriddedPerm]):
         workable: bool = True,
     ):
         self.partition = tuple(sorted(tuple(sorted(p)) for p in partition))
-        super().__init__(ignore_parent=ignore_parent, workable=workable)
+        inferrable = any(interleaving_rows_and_cols(self.partition))
+        super().__init__(
+            ignore_parent=ignore_parent, workable=workable, inferrable=inferrable
+        )
 
     def decomposition_function(self, tiling: Tiling) -> Tuple[Tiling, ...]:
         return tuple(tiling.sub_tiling(cells) for cells in self.partition)
@@ -60,9 +63,7 @@ class FactorStrategy(CartesianProductStrategy[Tiling, GriddedPerm]):
         ):
             for idx, child in enumerate(children):
                 # TODO: consider skew/sum
-                new_assumption = child.forward_map_assumption(
-                    assumption, check_avoidance=False
-                )
+                new_assumption = child.forward_map_assumption(assumption)
                 if new_assumption.gps:
                     child_var = child.get_parameter(new_assumption)
                     extra_parameters[idx][parent_var] = child_var
@@ -214,27 +215,6 @@ class Interleaving(CartesianProduct[Tiling, GriddedPerm]):
         super().__init__(parent, children, extra_parameters)
         self.interleaving_parameters = tuple(interleaving_parameters)
 
-    @classmethod
-    def untracked(
-        cls,
-        parent: Tiling,
-        children: Iterable[Tiling],
-        extra_parameters: Tuple[Dict[str, str], ...],
-    ) -> "Interleaving":
-        res = cls(parent, children, extra_parameters, [])
-
-        def f(*args, **kwargs):
-            raise NotImplementedError(
-                "enumeration for untracked interleaving factors is not implemented"
-            )
-
-        # Overwriting methods is bad practice, but useful here!
-        res.get_recurrence = f  # type: ignore
-        res.get_equation = f  # type: ignore
-        res.get_sub_objects = f  # type: ignore
-        res.random_sample_sub_objects = f  # type: ignore
-        return res
-
     @staticmethod
     def is_equivalence() -> bool:
         return False
@@ -270,14 +250,6 @@ class Interleaving(CartesianProduct[Tiling, GriddedPerm]):
     ):
         raise NotImplementedError
 
-    @staticmethod
-    def get_eq_symbol() -> str:
-        return "="
-
-    @staticmethod
-    def get_op_symbol() -> str:
-        return "*"
-
 
 class FactorWithInterleavingStrategy(FactorStrategy):
     def formal_step(self) -> str:
@@ -292,9 +264,7 @@ class FactorWithInterleavingStrategy(FactorStrategy):
             interleaving_parameters = self.interleaving_parameters(tiling)
         except ValueError:
             # must be untracked
-            return Interleaving.untracked(
-                tiling, children, self.extra_parameters(tiling, children)
-            )
+            raise NotImplementedError("The interleaving factor was not tracked.")
         return Interleaving(
             tiling,
             children,
@@ -346,6 +316,14 @@ class FactorWithInterleavingStrategy(FactorStrategy):
     ) -> Tuple[GriddedPerm, ...]:
         raise NotImplementedError
 
+    @staticmethod
+    def get_eq_symbol() -> str:
+        return "="
+
+    @staticmethod
+    def get_op_symbol() -> str:
+        return "*"
+
     def __repr__(self) -> str:
         return (
             self.__class__.__name__
@@ -367,12 +345,7 @@ class FactorWithMonotoneInterleavingStrategy(FactorWithInterleavingStrategy):
             interleaving_parameters = self.interleaving_parameters(tiling)
         except ValueError:
             # must be untracked
-            return cast(
-                MonotoneInterleaving,
-                MonotoneInterleaving.untracked(
-                    tiling, children, self.extra_parameters(tiling, children)
-                ),
-            )
+            raise NotImplementedError("The monotone interleaving was not tracked.")
         return MonotoneInterleaving(
             tiling,
             children,